+++ features/admin.claims.v1/components/edit/local-claim/edit-basic-details-local-claims.tsx
@@ -50,6 +50,7 @@ import {
     Claim,
     ClaimDataType,
     ClaimDialect,
+    ClaimInputFormat,
     ExternalClaim,
     ProfileSchemaInterface,
     SharedProfileValueResolvingMethod,
@@ -142,6 +143,8 @@ export const EditBasicDetailsLocalClaims: FunctionComponent<EditBasicDetailsLoca
     const [ hasMapping, setHasMapping ] = useState<boolean>(false);
     const [ mappingChecked, setMappingChecked ] = useState<boolean>(false);
     const [ dataType, setDataType ] = useState<string>(claim?.dataType || "");
+    const [ inputType, setInputType ] = useState<string>(claim?.inputFormat?.inputType || ClaimInputFormat.TEXT_INPUT);
+    const [ multiValued, setMultiValued ] = useState<boolean>(claim?.multiValued || false);
     const [ subAttributes, setSubAttributes ] = useState<string[]>([]);
     const [ canonicalValues, setCanonicalValues ] = useState<KeyValue[]>();
 
@@ -303,6 +306,7 @@ export const EditBasicDetailsLocalClaims: FunctionComponent<EditBasicDetailsLoca
             setDataType(claim?.dataType || ClaimDataType.STRING);
         }
 
+        setInputType(claim?.inputFormat?.inputType || ClaimInputFormat.TEXT_INPUT);
         setIsConsoleRequired(claim?.profiles?.console?.required ?? claim?.required);
         setIsEndUserRequired(claim?.profiles?.endUser?.required ?? claim?.required);
         setIsSelfRegistrationRequired(claim?.profiles?.selfRegistration?.required ?? claim?.required);
@@ -677,6 +681,7 @@ export const EditBasicDetailsLocalClaims: FunctionComponent<EditBasicDetailsLoca
                     : claim?.displayName,
                 displayOrder: attributeConfig.editAttributes.getDisplayOrder(
                     claim.displayOrder, values.displayOrder?.toString()),
+                inputFormat: inputType ? { inputType: inputType } : claim?.inputFormat,
                 multiValued: values?.multiValued !== undefined
                     ? !!values.multiValued : claim?.multiValued,
                 properties: claim?.properties,
@@ -725,6 +730,7 @@ export const EditBasicDetailsLocalClaims: FunctionComponent<EditBasicDetailsLoca
                     : claim?.displayName,
                 displayOrder: attributeConfig.editAttributes.getDisplayOrder(
                     claim.displayOrder, values.displayOrder?.toString()),
+                inputFormat: inputType ? { inputType: inputType } : claim?.inputFormat,
                 multiValued: values?.multiValued !== undefined
                     ? !!values.multiValued : claim?.multiValued,
                 profiles: {
@@ -891,6 +897,93 @@ export const EditBasicDetailsLocalClaims: FunctionComponent<EditBasicDetailsLoca
                 </TableCell>
             </TableRow>
         );
+    };
+
+    const setDefaultInputTypeForDataType = (dataType: ClaimDataType, multiValued: boolean): void => {
+
+        switch (dataType) {
+            case ClaimDataType.OPTIONS:
+                setInputType(multiValued ? ClaimInputFormat.MULTI_SELECT_DROPDOWN : ClaimInputFormat.DROPDOWN);
+
+                break;
+            case ClaimDataType.BOOLEAN:
+                setInputType(ClaimInputFormat.CHECKBOX);
+
+                break;
+            default:
+                setInputType(ClaimInputFormat.TEXT_INPUT);
+        }
+    };
+
+    const resolveInputFormatOptions = (dataType: ClaimDataType, multiValued: boolean): DropDownItemInterface[] => {
+
+        const textInputOption: DropDownItemInterface[] = [
+            {
+                text: t("claims:local.forms.inputFormat.options.textInput"),
+                value: ClaimInputFormat.TEXT_INPUT
+            }
+        ];
+
+        if (dataType === ClaimDataType.DATE_TIME) {
+            return [
+                ...textInputOption,
+                {
+                    text: t("claims:local.forms.inputFormat.options.datePicker"),
+                    value: ClaimInputFormat.DATE_PICKER
+                }
+            ];
+        }
+
+        if (dataType === ClaimDataType.OPTIONS) {
+            if (multiValued) {
+                return [
+                    {
+                        text: t("claims:local.forms.inputFormat.options.multiSelectDropdown"),
+                        value: ClaimInputFormat.MULTI_SELECT_DROPDOWN
+                    },
+                    {
+                        text: t("claims:local.forms.inputFormat.options.checkBoxGroup"),
+                        value: ClaimInputFormat.CHECKBOX_GROUP
+                    }
+                ];
+            }
+
+            return [
+                {
+                    text: t("claims:local.forms.inputFormat.options.radioGroup"),
+                    value: ClaimInputFormat.RADIO_GROUP
+                },
+                {
+                    text: t("claims:local.forms.inputFormat.options.dropdown"),
+                    value: ClaimInputFormat.DROPDOWN
+                }
+            ];
+        }
+
+        if (dataType === ClaimDataType.INTEGER) {
+            return [
+                ...textInputOption,
+                {
+                    text: t("claims:local.forms.inputFormat.options.numberInput"),
+                    value: ClaimInputFormat.NUMBER_INPUT
+                }
+            ];
+        }
+
+        if (dataType === ClaimDataType.BOOLEAN) {
+            return [
+                {
+                    text: t("claims:local.forms.inputFormat.options.checkbox"),
+                    value: ClaimInputFormat.CHECKBOX
+                },
+                {
+                    text: t("claims:local.forms.inputFormat.options.toggle"),
+                    value: ClaimInputFormat.TOGGLE
+                }
+            ];
+        }
+
+        return textInputOption;
     };
 
     const resolveAttributeRequiredRow = (): ReactElement => {
@@ -1124,6 +1217,7 @@ export const EditBasicDetailsLocalClaims: FunctionComponent<EditBasicDetailsLoca
                             setSubAttributes([]);
                             setCanonicalValues([]);
                             setDataType(data.value);
+                            setDefaultInputTypeForDataType(data.value as ClaimDataType, multiValued);
                         } }
                     />
 
@@ -1144,6 +1238,7 @@ export const EditBasicDetailsLocalClaims: FunctionComponent<EditBasicDetailsLoca
                                     }
                                 } }
                                 data-testid={ `${testId}-form-sub-attributes-dropdown` }
+                                disabled={ isSubOrganization() || isSystemClaim || isReadOnly }
                                 search
                             />
                             <div className="sub-attribute-list">
@@ -1152,21 +1247,23 @@ export const EditBasicDetailsLocalClaims: FunctionComponent<EditBasicDetailsLoca
                                         className="sub-attribute-row"
                                         key={ index }>
                                         <span>{ attribute }</span>
-                                        <IconButton
-                                            className="sub-attribute-delete-btn"
-                                            disabled={ isReadOnly }
-                                            onClick={ () => {
-                                                setSubAttributes(
-                                                    subAttributes.filter(
-                                                        (item: string) => item !== attribute
-                                                    )
-                                                );
-                                            } }
-                                            data-componentid={ `${testId}-delete-sub-attribute-${index}` }
-                                            style={ { marginLeft: "auto" } }
-                                        >
-                                            <TrashIcon />
-                                        </IconButton>
+                                        { !(isSubOrganization() || isSystemClaim || isReadOnly) && (
+                                            <IconButton
+                                                className="sub-attribute-delete-btn"
+                                                disabled={ isReadOnly }
+                                                onClick={ () => {
+                                                    setSubAttributes(
+                                                        subAttributes.filter(
+                                                            (item: string) => item !== attribute
+                                                        )
+                                                    );
+                                                } }
+                                                data-componentid={ `${testId}-delete-sub-attribute-${index}` }
+                                                style={ { marginLeft: "auto" } }
+                                            >
+                                                <TrashIcon />
+                                            </IconButton>
+                                        ) }
                                     </div>
                                 )) }
                             </div>
@@ -1191,7 +1288,7 @@ export const EditBasicDetailsLocalClaims: FunctionComponent<EditBasicDetailsLoca
                                         ({ key: item.key, value: item.value })));
                                 } }
                                 data-testid={ `${testId}-form-canonical-values-dynamic-field` }
-                                readOnly={ isReadOnly }
+                                readOnly={ isSubOrganization() || isReadOnly }
                             />
                         </>
                     ) }
@@ -1203,11 +1300,33 @@ export const EditBasicDetailsLocalClaims: FunctionComponent<EditBasicDetailsLoca
                             label={ t("claims:local.forms.multiValued.label") }
                             required={ false }
                             defaultValue={ claim?.multiValued }
-                            data-testid={ `${testId}-form-multi-valued-input` }
+                            data-componentid={ `${testId}-form-multi-valued-input` }
                             hint={ isSystemClaim
                                 ? t("claims:local.forms.multiValuedSystemClaimHint")
                                 : t("claims:local.forms.multiValuedHint") }
                             readOnly={ isSubOrganization() || isSystemClaim || isReadOnly }
+                            listen={ (checked: boolean) => {
+                                setMultiValued(checked);
+                                setDefaultInputTypeForDataType(dataType as ClaimDataType, checked);
+                            } }
+                        />
+                    ) }
+                    { dataType !== ClaimDataType.COMPLEX && (
+                        <Field.Dropdown
+                            ariaLabel={ t("claims:local.forms.inputFormat.label") }
+                            name="inputFormat"
+                            label={ t("claims:local.forms.inputFormat.label") }
+                            data-componentid={ `${testId}-form-input-format-input` }
+                            hint={ t("claims:local.forms.inputFormat.hint") }
+                            disabled={ isSubOrganization() || isReadOnly }
+                            options={ resolveInputFormatOptions(dataType as ClaimDataType, multiValued) }
+                            value={ inputType }
+                            onChange={ (
+                                event: React.SyntheticEvent<HTMLElement, Event>,
+                                data: { value: string }
+                            ) => {
+                                setInputType(data.value);
+                            } }
                         />
                     ) }
                     { !attributeConfig.localAttributes.createWizard.showRegularExpression && !hideSpecialClaims